back-end/wearhouse/requirements.txt
coverage==4.5.4
Django==2.2.6
pylint==2.4.3
mysqlclient==1.4.4
Pillow==6.2.1
django-storages==1.7.2
coveralls==1.8.2
requests==2.22.0
